Add standalone import page
